+++ apps/chaingraph-frontend/src/store/nodes/stores.ts
@@ -34,7 +34,7 @@ import { $trpcClient } from '../trpc/store'
 import { LOCAL_NODE_UI_DEBOUNCE_MS, NODE_POSITION_DEBOUNCE_MS, NODE_UI_DEBOUNCE_MS } from './constants'
 import { accumulateAndSample } from './operators/accumulate-and-sample'
 import { positionInterpolator } from './position-interpolation-advanced'
-import './interpolation-init'
+// import './computed'
 
 // EVENTS