+++ packages/badai-api/src/gql/client/fragment-masking.ts
@@ -7,13 +7,9 @@
  */
 
 /* eslint-disable */
-import {
-  DocumentTypeDecoration,
-  ResultOf,
-  TypedDocumentNode,
-} from '@graphql-typed-document-node/core'
-import { FragmentDefinitionNode } from 'graphql'
-import { Incremental } from './graphql'
+import { ResultOf, DocumentTypeDecoration, TypedDocumentNode } from '@graphql-typed-document-node/core';
+import { FragmentDefinitionNode } from 'graphql';
+import { Incremental } from './graphql';
 
 
 export type FragmentType<TDocumentType extends DocumentTypeDecoration<any, any>> = TDocumentType extends DocumentTypeDecoration<